refactor: update build process and optimize protobuf usage in Node.js client
