+++ node/src/GlideClusterClient.ts
@@ -2,9 +2,8 @@
  * Copyright Valkey GLIDE Project Contributors - SPDX Identifier: Apache-2.0
  */
 
-import { ClusterScanCursor, Script } from "./native";
 import * as net from "net";
-import { Writer } from "protobufjs";
+import { Writer } from "protobufjs/minimal";
 import {
     AdvancedBaseClientConfiguration,
     BaseClient,
@@ -64,7 +63,9 @@ import {
     createTime,
     createUnWatch,
 } from "./Commands";
-import { command_request, connection_request } from "./ProtobufMessage";
+import { ClusterScanCursor, Script } from "./native";
+import { command_request, connection_request } from "../build-ts/ProtobufMessage";
+;
 
 /** An extension to command option types with {@link Routes}. */
 export interface RouteOption {
